# 8358693: [leyden] Bootstrapping circularity leads to never starting preload threads