@shipilev 2025-06-05
> The effect on current premain is less pronounced, because I think there are more regressions. I am chasing those down as well.

Now that https://github.com/openjdk/leyden/pull/76 has integrated, we have the performance improvements from this PR on current premain too. See:

```
# Current premain
 Time (mean ± σ):     370.1 ms ±   5.9 ms    [User: 909.2 ms, System: 137.2 ms]
 Range (min … max):   363.8 ms … 389.9 ms    30 runs

# This PR
 Time (mean ± σ):     329.1 ms ±   5.0 ms    [User: 667.7 ms, System: 104.6 ms]
 Range (min … max):   320.9 ms … 349.9 ms    30 runs
```