@shipilev 2025-06-05
> Please update `AOTCodeCache::is_code_load_thread_on()` instead.

That also works, updated there.