@shipilev 2025-06-05
I would expect bots are choked up with all this JDK 25 fun? :)