@shipilev 2025-06-05
/integrate